Split fluff_lint module into fluff/ package
